angular: add 'getJobs()',url_resume_download to resume.svc
