+++ src/styles.css
@@ -12,6 +12,10 @@ body {
   padding-bottom: 5px;
 }
 
+.pbl {
+  padding-bottom: 15px;
+}
+
 .pts {
   padding-top: 5px;
 }
